Rename Python binding module for clarity
